refactor: move Sidebar to global authenticated layout
- Create (authenticated) route group with shared layout
- Move Sidebar from components/board to components/layout
- Wrap authenticated pages with OrgGuard in layout
- Simplify TaskBoard and TaskDetailsPage by removing Sidebar
- Delete old (board) and task directories
